taluka and gram panchayat api done

+++ src/controller/DistrictController.ts
@@ -1,13 +1,21 @@
 import { Logger } from '../logger/Logger';
-import {createDistrict, deleteDistrict, getAllDistrictForDDL, getAllDistricts, getDistrict, updateDistrict} from '../models/District';
+import {createDistrict, deleteDistrict, findDistrict, getAllDistrictForDDL, getAllDistricts, getDistrict, getDistrictCount, updateDistrict} from '../models/District';
 import { ApiResponse } from '../utils/ApiResponse';
 const logger = new Logger().logger;
 export class DistrictController {
     static async createDisrict(req, res, next) {
         try {
             let { name } = req.body;
-            let district: any = await createDistrict(name);
-            let response = ApiResponse.successResponse(res, district.name+" District created successfully");
+             let params = {
+                searchText: name
+            }
+            let avialableDistrict = await findDistrict(params);
+            if(avialableDistrict.length == 0){
+                let district: any = await createDistrict(name);
+                var response = ApiResponse.successResponse(res, district.name+" District created successfully");
+            }else{
+                var response = ApiResponse.successResponse(res, name+" District alraedy exist");
+            }
             return response;
         } catch (error) {
             // logger.info("action:User/login", { message: `User : ${user?.username} logged in successfully at ${new Date().toISOString()}` });
@@ -29,7 +37,8 @@ export class DistrictController {
             }
 
             let districts = await getAllDistricts(params);
-            let response = ApiResponse.successResponseWithData(res, "Districts fetched successfully", districts);
+            let totalCount = (searchText == '') ? await getDistrictCount() : Object.keys(districts).length;
+            let response = ApiResponse.successResponseWithData(res, "Districts fetched successfully", {districts, totalCount: totalCount});
             return response;
         } catch (error) {
             let response = ApiResponse.ErrorResponse(res, "Failed to fetch Districts", error);
@@ -40,8 +49,11 @@ export class DistrictController {
     static async getDistrict(req, res, next) {
         try {
             let data = req.body;
-            console.log(data.id);
             let district = await getDistrict(parseInt(data.id));
+            if(district == null){
+                let response = ApiResponse.successResponseWithData(res, "District not found", []);
+                return response;
+            }
             let response = ApiResponse.successResponseWithData(res, "District fetched successfully", district);
             return response;
         } catch (error) {
@@ -65,6 +77,11 @@ export class DistrictController {
     static async deleteDistrict(req, res, next) {
         try {
             let data = req.body;
+            let existDistrict = await getDistrict(parseInt(data.id));
+            if(existDistrict == null){
+                let response = ApiResponse.successResponse(res, "Sorry we'r unable to find this district for delete operation");
+                return response;
+            }
             let district = await deleteDistrict(parseInt(data.id));
             let response = ApiResponse.successResponseWithData(res, "District deleted successfully", district);
             return response;
