+++ src/controller/TalukaController.ts
@@ -1,12 +1,20 @@
 
-// import {createDistrict, deleteDistrict, getAllDistricts, getDistrict, updateDistrict} from '../models/District';
-import { createTaluka, deleteTaluka, getAllTalukas, getTaluka, getTalukaCount, updateTaluka } from '../models/Taluka';
+import { createTaluka, deleteTaluka, findTaluka, getAllTalukas, getTaluka, getTalukaCount, updateTaluka } from '../models/Taluka';
 import { ApiResponse } from '../utils/ApiResponse';
 export class TalukaController {
 
     static async createTaluka(req, res, next) {
         try {
             let { name, district_id } = req.body;
+            let params = {
+                id: district_id,
+                searchText: name
+            }
+            let availableTaluka = await findTaluka(params);
+            if(availableTaluka.length > 0){
+                let response = ApiResponse.successResponse(res, name+ " Taluka already exists");
+                return response;
+            }
             let taluka: any = await createTaluka(name, district_id);
             let response = ApiResponse.successResponse(res, taluka.name + " Taluka created successfully");
             return response;
@@ -18,7 +26,7 @@ export class TalukaController {
 
     static async getAllTaluka(req, res, next) {
         try {
-            let offset:any = process.env.PAGE_OFFSET || 0;
+             let offset:number = parseInt(process.env.PAGE_OFFSET) || 0;
             let pageNumber = (req.body?.pageNumber ?? 1) - 1;
             let searchText = req.body?.searchText ?? "";
             let limit = (pageNumber != 0) ? pageNumber * offset : pageNumber;
@@ -41,6 +49,10 @@ export class TalukaController {
         try {
             let data = req.body;
             let taluka = await getTaluka(parseInt(data.id));
+            if(taluka == null){
+                let response = ApiResponse.successResponseWithData(res, "Taluka not found", []);
+                return response;
+            }
             let response = ApiResponse.successResponseWithData(res, "Taluka fetched successfully", taluka);
             return response;
         } catch (error) {
@@ -64,6 +76,11 @@ export class TalukaController {
     static async deleteTaluka(req, res, next) {
         try {
             let data = req.body;
+            let existTaluka = await getTaluka(parseInt(data.id));
+            if(existTaluka == null){
+                let response = ApiResponse.successResponse(res, "Sorry we'r unable to find this taluka for delete operation");
+                return response;
+            }
             let taluka = await deleteTaluka(parseInt(data.id));
             let response = ApiResponse.successResponseWithData(res, "Taluka deleted successfully", taluka);
             return response;
